Removed candidate check from the individual scaler
At some prior point, it was necessary to check that there were
applicable scaling candidates in the individual strategy during
the actual up/down scale processes.

This check now happens well before the strategy actually gets
called, so it is no longer possible for the strategy to
receive an empty list of candidates.